server/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
beautifulsoup4
fpdf2
